# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MilestoneContentDifficulty = typeof MilestoneContentDifficulty[keyof typeof MilestoneContentDifficulty];


export const MilestoneContentDifficulty = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
} as const;

export interface MilestoneContent {
  /** @maxLength 150 */
  name: string;
  /** @maxLength 1000 */
  description: string;
  /** @maxLength 500 */
  command: string;
  /** @maxLength 2000 */
  handSignal: string;
  /** @maxLength 12000 */
  instructions: string;
  /** @maxLength 100 */
  category: string;
  difficulty: MilestoneContentDifficulty;
  /** @maxLength 500 */
  videoPath: string;
}

export type MilestoneTemplate = MilestoneContent & {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
};

export interface PathStep {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  templateId: string;
  customization: MilestoneContent | null;
}

export interface TrainingProgram {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /** @maxLength 2000 */
  description: string;
  /** @maxItems 200 */
  steps: PathStep[];
}

export interface DogProfile {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /** @maxLength 150 */
  breed: string;
  /** @maxLength 150 */
  trainingType: string;
  /** @maxLength 10 */
  birthDate: string;
  /** @maxLength 150 */
  handlerName: string;
  /** @maxLength 500 */
  photo: string;
}

export type DogAssignmentProgress = {[key: string]: 'not-started' | 'in-progress' | 'completed'};

export interface DogAssignment {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  dogId: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  pathId: string;
  progress: DogAssignmentProgress;
}

export type TrainingWorkspaceVersion = typeof TrainingWorkspaceVersion[keyof typeof TrainingWorkspaceVersion];


export const TrainingWorkspaceVersion = {
  NUMBER_2: 2,
} as const;

export interface TrainingWorkspace {
  version: TrainingWorkspaceVersion;
  /** @maxItems 300 */
  dogs: DogProfile[];
  /** @maxItems 300 */
  paths: TrainingProgram[];
  /** @maxItems 1000 */
  milestones: MilestoneTemplate[];
  /** @maxItems 1000 */
  assignments: DogAssignment[];
}

export interface WorkspaceInput {
  state: TrainingWorkspace;
  /** @minimum 0 */
  revision: number;
}

export interface WorkspaceResponse {
  state: TrainingWorkspace | null;
  revision: number;
  /** @nullable */
  updatedAt: string | null;
}

export interface HandlerShareInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  assignmentId: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  handlerName: string;
}

export interface HandlerShare {
  id: string;
  assignmentId: string;
  handlerName: string;
}

export type HandlerMilestoneStatus = typeof HandlerMilestoneStatus[keyof typeof HandlerMilestoneStatus];


export const HandlerMilestoneStatus = {
  'not-started': 'not-started',
  'in-progress': 'in-progress',
  completed: 'completed',
} as const;

export type HandlerMilestone = MilestoneContent & {
  id: string;
  status: HandlerMilestoneStatus;
};

export interface HandlerPath {
  dog: DogProfile;
  pathName: string;
  pathDescription: string;
  handlerName: string;
  milestones: HandlerMilestone[];
  updatedAt: string;
}

export type UploadInputContentType = typeof UploadInputContentType[keyof typeof UploadInputContentType];


export const UploadInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'video/mp4': 'video/mp4',
  'video/webm': 'video/webm',
  'video/quicktime': 'video/quicktime',
} as const;

export interface UploadInput {
  /** @maxLength 250 */
  name: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  size: number;
  contentType: UploadInputContentType;
}

export interface UploadTicket {
  uploadURL: string;
  objectPath: string;
}

export interface MediaInput {
  /** @maxLength 200 */
  objectPath: string;
}

export interface MediaResult {
  objectPath: string;
}