# lib/api-zod/src/generated/types/dogProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DogProfile {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  name: string;
  /** @maxLength 150 */
  breed: string;
  /** @maxLength 150 */
  trainingType: string;
  /** @maxLength 10 */
  birthDate: string;
  /** @maxLength 150 */
  handlerName: string;
  /** @maxLength 500 */
  photo: string;
}