# lib/api-zod/src/generated/types/handlerMilestone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HandlerMilestoneStatus } from './handlerMilestoneStatus';
import type { MilestoneContent } from './milestoneContent';

export type HandlerMilestone = MilestoneContent & {
  id: string;
  status: HandlerMilestoneStatus;
};