# lib/api-zod/src/generated/types/handlerPath.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DogProfile } from './dogProfile';
import type { HandlerMilestone } from './handlerMilestone';

export interface HandlerPath {
  dog: DogProfile;
  pathName: string;
  pathDescription: string;
  handlerName: string;
  milestones: HandlerMilestone[];
  updatedAt: string;
}