# lib/api-zod/src/generated/types/handlerShare.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HandlerShare {
  id: string;
  assignmentId: string;
  handlerName: string;
}