# lib/api-zod/src/generated/types/handlerShareInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HandlerShareInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  assignmentId: string;
  /**
     * @minLength 1
     * @maxLength 150
     */
  handlerName: string;
}