# lib/api-zod/src/generated/types/milestoneTemplate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MilestoneContent } from './milestoneContent';

export type MilestoneTemplate = MilestoneContent & {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
};