# lib/api-zod/src/generated/types/pathStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MilestoneContent } from './milestoneContent';

export interface PathStep {
  /**
     * @minLength 1
     * @maxLength 100
     */
  id: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  templateId: string;
  customization: MilestoneContent | null;
}