# lib/api-zod/src/generated/types/trainingWorkspace.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DogAssignment } from './dogAssignment';
import type { DogProfile } from './dogProfile';
import type { MilestoneTemplate } from './milestoneTemplate';
import type { TrainingProgram } from './trainingProgram';
import type { TrainingWorkspaceVersion } from './trainingWorkspaceVersion';

export interface TrainingWorkspace {
  version: TrainingWorkspaceVersion;
  /** @maxItems 300 */
  dogs: DogProfile[];
  /** @maxItems 300 */
  paths: TrainingProgram[];
  /** @maxItems 1000 */
  milestones: MilestoneTemplate[];
  /** @maxItems 1000 */
  assignments: DogAssignment[];
}