# lib/api-zod/src/generated/types/workspaceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrainingWorkspace } from './trainingWorkspace';

export interface WorkspaceResponse {
  state: TrainingWorkspace | null;
  revision: number;
  /** @nullable */
  updatedAt: string | null;
}